refactor: move split[Url]Path() & test
