+++ src/parsers/Jenkins.spec.ts
@@ -13,22 +13,6 @@ function testParseLink(html: string, url: string): Link | null {
     return actual;
 }
 
-test("splitUrlPath with deep URL", () => {
-    const input = "http://localhost:8080/job/Project/job/Repository/job/Branch/";
-    const cut = new Jenkins();
-
-    const actual = cut.splitUrlPath(input);
-
-    assert.equal(actual.length, 7);
-    assert.equal(actual[0], "");
-    assert.equal(actual[1], "job");
-    assert.equal(actual[2], "Project");
-    assert.equal(actual[3], "job");
-    assert.equal(actual[4], "Repository");
-    assert.equal(actual[5], "job");
-    assert.equal(actual[6], "Branch");
-});
-
 test("job page", () => {
     const html = `
 <html>
@@ -235,7 +219,10 @@ test("run's console output", () => {
         actual?.destination,
         "http://localhost:8080/job/Project/job/Repository/job/Branch/1/console"
     );
-    assert.equal(actual?.text, "Project » Repository » Branch #1 » Console Output");
+    assert.equal(
+        actual?.text,
+        "Project » Repository » Branch #1 » Console Output"
+    );
 });
 
 test("run page, version 2.361.4", () => {
